+++ resources/js/Components/VideoPlayer/VideoControls/VideoControlsFullPage.vue
@@ -14,8 +14,6 @@
 
         <div class="videoControls">
 
-
-
             <VideoControlsButtons />
 
         </div>